refactor: Have changed Loader from normal css to modular
