--- src/components/Loader.tsx
@@ -1,11 +1,0 @@
-import './Loader.css';
-
-const Loader = () => {
-  return (
-    <div className="loader-container">
-      <div className="spinner" />
-    </div>
-  );
-};
-
-export default Loader;
