+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,11 @@
+import styles from "./Loader.module.css"
+
+const Loader = () => {
+  return (
+    <div className={styles.loaderContainer}>
+      <div className={styles.spinner} />
+    </div>
+  );
+};
+
+export default Loader;
